Inventory Save and Load Completed
Implemented Collectable for the Player Inventory not perfect but its a
start.
Moved collectable items into the Collectable Subclass
Inventory is now an ArrayBuffer[Option[Collectable]]
The Item factory will still be able to create Collectable Items but they
are returned as Item instead of Collectable which is needed for the
Inventory

equalCollectable method is used by the inventory to see if a collectable
exists in the inventory already or if it needs to add it.
CastleLoading State uses the sharedStateData to take an XML file and
load it into a Castle.
Modified MapLoad to use the Collectable subclass for player inventory