Created Castle Select Screen
The castle select screen will get all .map files in the maps folder of
the working directory
It loads each of them to get the metadata then displays it in a
scrollable list of items

Enter loads the game screen while 'E' Loads the editor screen

Needs visual cleanup but this is a functional version

CastleLoading was modified to allow you to choose what screen to load
after it is finished
TitleScreen's input handling was changed because of a latient bug in the
Enter being cached since it was not polling isKeyPressed the MenuScreen
would see that Enter was pressed since the last time isKeyPressed was
checked.

Fixed the CastleCreateScreen's castle loading.
SharedStateData.loadedCastle must be assigned to a temp variable since
after it has been retrieved once it will be reset to null.