Threaded Saving
MapSave will now clone the map using the cloning library. This allows us
to have a threadsafe (hopefully) map to save.

Used a future block surrounding the saving portion of code for the non
editor. Everything inside of this block is ran as its own thread.

saveDone allows you to check if the future block is done from another
part of the code and block until it is using saveDone.apply()